requirements/base.txt
Django==3.0.3     
django-allauth==0.41.0    
django-discover-runner==1.0       
django-environ==0.4.5
django-filer==1.7.0     
django-filter==2.2.0     
django-phonenumber-field==4.0.0 
phonenumbers==8.11.5   
django-rest-auth==0.9.5     
django-role-permissions==2.2.1     
djangorestframework==3.10.3    
django-cors-headers==3.2.1 
django-extra-fields==0.9        
django-taggit==1.2.0   
django-taggit-serializer==0.1
ffmpeg==1.4  
Pillow==7.0.0  
easy-thumbnails==2.7 
psycopg2-binary==2.8.4 
six==1.14.0